requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.1.1
django-debug-toolbar==2.2
-e git+https://github.com/dellsystem/django-language-field@a7f23fb4547e5633c54f71914091f9f8cf6c34ba#egg=django-language-field
Fabric3==1.14.post1
-e git+https://github.com/dellsystem/goodreads@23a693fe52c2f4fb6daf6e65b7277102de288943#egg=goodreads
fakeredis==1.4.3
idna==2.8
Markdown==3.0.1
MarkdownSuperscript==2.1.1
-e git+https://github.com/dellsystem/merriam-webster-api@d49a108bb0e6c7d1e0a068e56055c7a372115431#egg=Merriam_Webster_API
nose==1.3.7
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.3
rauth==0.7.3
redis==3.5.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
terminaltables==3.1.0
-e git+https://github.com/dellsystem/translate-term.git@e1a585826d64cb16c83d783a16b0e84d848bfc44#egg=translate_term
urllib3==1.25.7
xmltodict==0.12.0
